Moving eye around start
IAN!!! it looks like a bad merge got in. I kept the "Just in case" diff
in but left the HEAD around too just commented out. I tried some stuff
to manipulate the eye position and this seems to be a decent solution
but I havent put much time in yet,