trying to make things work
I have functions in place but I am obviously doing something wrong. if
you press the "1" key you will get the positions (at least I think) of
the missiles and the warbird  and when you press space bar you will fire
a missile but I dont see it on screen. I also noticed that one of the
missiles seems to be following one of the planets around. I cant explain
that and the other seems to be no where in sight. Help mw
Ian-wan-kenobi, you are my only hope.